+++ lib/overlay_event.dart
@@ -1,0 +1,6 @@
+class OverlayEvent {
+  final String event;
+  final dynamic data;
+
+  OverlayEvent(this.event, this.data);
+}