+++ lib/screen_event.dart
@@ -1,6 +1,0 @@
-class ScreenEvent {
-  final String event;
-  final dynamic data;
-
-  ScreenEvent(this.event, this.data);
-}