+++ lib/screen_overlay.dart
@@ -1,6 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:screens/screen_widget.dart';
-
-abstract class ScreenOverlay {
-  Widget build(ScreenWidget screenWidget, dynamic data);
-}